Update laravel homestead master scripts as of.

+++ scripts/homestead.rb
@@ -1,5 +1,5 @@
 class Homestead
-  def Homestead.configure(config, settings, hostsupdater)
+  def Homestead.configure(config, settings)
     # Set The VM Provider
     ENV['VAGRANT_DEFAULT_PROVIDER'] = settings["provider"] ||= "virtualbox"
 
@@ -10,8 +10,8 @@ class Homestead
     config.ssh.shell = "bash -c 'BASH_ENV=/etc/profile exec bash'"
 
     # Configure The Box
-    config.vm.box = settings["box"] ||= "laravel/homestead-7"
-    config.vm.box_version = settings["version"] ||= ">= 0"
+    config.vm.box = settings["box"] ||= "laravel/homestead"
+    config.vm.box_version = settings["version"] ||= ">= 0.4.0"
     config.vm.hostname = settings["hostname"] ||= "homestead"
 
     # Configure A Private Network IP
@@ -26,7 +26,7 @@ class Homestead
 
     # Configure A Few VirtualBox Settings
     config.vm.provider "virtualbox" do |vb|
-      vb.name = 'homestead'
+      vb.name = settings["name"] ||= "homestead-7"
       vb.customize ["modifyvm", :id, "--memory", settings["memory"] ||= "2048"]
       vb.customize ["modifyvm", :id, "--cpus", settings["cpus"] ||= "1"]
       vb.customize ["modifyvm", :id, "--natdnsproxy1", "on"]
@@ -37,7 +37,7 @@ class Homestead
     # Configure A Few VMware Settings
     ["vmware_fusion", "vmware_workstation"].each do |vmware|
       config.vm.provider vmware do |v|
-        v.vmx["displayName"] = "homestead"
+        v.vmx["displayName"] = settings["name"] ||= "homestead-7"
         v.vmx["memsize"] = settings["memory"] ||= 2048
         v.vmx["numvcpus"] = settings["cpus"] ||= 1
         v.vmx["guestOS"] = "ubuntu-64"
@@ -47,7 +47,6 @@ class Homestead
     # Configure A Few Parallels Settings
     config.vm.provider "parallels" do |v|
       v.update_guest_tools = true
-      v.optimize_power_consumption = false
       v.memory = settings["memory"] ||= 2048
       v.cpus = settings["cpus"] ||= 1
     end
@@ -90,9 +89,11 @@ class Homestead
     }
 
     # Use Default Port Forwarding Unless Overridden
-    default_ports.each do |guest, host|
-      unless settings["ports"].any? { |mapping| mapping["guest"] == guest }
-        config.vm.network "forwarded_port", guest: guest, host: host, auto_correct: true
+    unless settings.has_key?("default_ports") && settings["default_ports"] == false
+      default_ports.each do |guest, host|
+        unless settings["ports"].any? { |mapping| mapping["guest"] == guest }
+          config.vm.network "forwarded_port", guest: guest, host: host, auto_correct: true
+        end
       end
     end
 
@@ -105,9 +106,11 @@ class Homestead
 
     # Configure The Public Key For SSH Access
     if settings.include? 'authorize'
-      config.vm.provision "shell" do |s|
-        s.inline = "echo $1 | grep -xq \"$1\" /home/vagrant/.ssh/authorized_keys || echo $1 | tee -a /home/vagrant/.ssh/authorized_keys"
-        s.args = [File.read(File.expand_path(settings["authorize"]))]
+      if File.exists? File.expand_path(settings["authorize"])
+        config.vm.provision "shell" do |s|
+          s.inline = "echo $1 | grep -xq \"$1\" /home/vagrant/.ssh/authorized_keys || echo $1 | tee -a /home/vagrant/.ssh/authorized_keys"
+          s.args = [File.read(File.expand_path(settings["authorize"]))]
+        end
       end
     end
 
@@ -122,16 +125,34 @@ class Homestead
       end
     end
 
+    # Copy User Files Over to VM
+    if settings.include? 'copy'
+      settings["copy"].each do |file|
+        config.vm.provision "file" do |f|
+          f.source = File.expand_path(file["from"])
+          f.destination = file["to"].chomp('/') + "/" + file["from"].split('/').last
+        end
+      end
+    end
+
     # Register All Of The Configured Shared Folders
     if settings.include? 'folders'
       settings["folders"].each do |folder|
         mount_opts = []
 
         if (folder["type"] == "nfs")
-            mount_opts = folder["mount_opts"] ? folder["mount_opts"] : ['actimeo=1']
+            mount_opts = folder["mount_options"] ? folder["mount_options"] : ['actimeo=1']
+        elsif (folder["type"] == "smb")
+            mount_opts = folder["mount_options"] ? folder["mount_options"] : ['vers=3.02', 'mfsymlinks']
         end
 
-        config.vm.synced_folder folder["map"], folder["to"], type: folder["type"] ||= nil, mount_options: mount_opts
+        # For b/w compatibility keep separate 'mount_opts', but merge with options
+        options = (folder["options"] || {}).merge({ mount_options: mount_opts })
+
+        # Double-splat (**) operator only works with symbol keys, so convert
+        options.keys.each{|k| options[k.to_sym] = options.delete(k) }
+
+        config.vm.synced_folder folder["map"], folder["to"], type: folder["type"] ||= nil, **options
       end
     end
 
@@ -140,20 +161,43 @@ class Homestead
         s.path = scriptDir + "/clear-nginx.sh"
     end
 
-    settings["sites"].each do |site|
-      type = site["type"] ||= "laravel"
+    if settings.include? 'sites'
+      settings["sites"].each do |site|
+        type = site["type"] ||= "laravel"
 
-      if (site.has_key?("hhvm") && site["hhvm"])
-        type = "hhvm"
+        if (site.has_key?("hhvm") && site["hhvm"])
+          type = "hhvm"
+        end
+
+        if (type == "symfony")
+          type = "symfony2"
+        end
+
+        config.vm.provision "shell" do |s|
+          s.path = scriptDir + "/serve-#{type}.sh"
+          s.args = [site["map"], site["to"], site["port"] ||= "80", site["ssl"] ||= "443"]
+        end
+
+        # Configure The Cron Schedule
+        if (site.has_key?("schedule"))
+          config.vm.provision "shell" do |s|
+            if (site["schedule"])
+              s.path = scriptDir + "/cron-schedule.sh"
+              s.args = [site["map"].tr('^A-Za-z0-9', ''), site["to"]]
+            else
+              s.inline = "rm -f /etc/cron.d/$1"
+              s.args = [site["map"].tr('^A-Za-z0-9', '')]
+            end
+          end
+        end
+
       end
+    end
 
-      if (type == "symfony")
-        type = "symfony2"
-      end
-
+    # Install MariaDB If Necessary
+    if settings.has_key?("mariadb") && settings["mariadb"]
       config.vm.provision "shell" do |s|
-        s.path = scriptDir + "/serve-#{type}.sh"
-        s.args = [site["map"], site["to"], site["port"] ||= "80", site["ssl"] ||= "443"]
+        s.path = scriptDir + "/install-maria.sh"
       end
     end
 
@@ -175,15 +219,6 @@ class Homestead
           s.args = [db]
         end
       end
-    end
-
-    # Add the aliases to the host machine
-    if hostsupdater
-      aliases = Array.new
-      settings["sites"].each do |site|
-        aliases.push(site["map"])
-      end
-      config.hostsupdater.aliases = aliases
     end
 
     # Configure All Of The Server Environment Variables
